requirements.txt
# Core packages needed for our notification service
flask==2.0.1
werkzeug==2.0.2
pika==1.2.0
retrying==1.3.3
pymongo==4.1.1
dnspython==2.2.1
python-dotenv==0.19.2

# Deployment related packages
gunicorn==20.1.0
prometheus-client==0.14.1 # For metrics
pytest==7.3.1 # For testing
coverage==7.2.5 # For test coverage
